Have a singular GUI component
