--- src/version.py
@@ -1,4 +1,0 @@
-VERSION = "0.0.11"
-
-def get_version():
-    return VERSION
